🌱 Remove legacy hub cookie mint helpers
Move legacy/v2 hub session-cookie minting and telemetry helpers into test-only code so production binaries no longer compile the forgery-capable HMAC lane while rejection tests keep their fixtures.\n\nFixes #5812\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
