+++ changelog.d/security-5812-dead-hub-cookie.md
@@ -1,0 +1,1 @@
+security: remove legacy hub session-cookie mint and verification helpers from the production binary while keeping rejection regression tests.
